+++ systems/__init__.py
@@ -1,6 +1,8 @@
+from .JOptSocAmB33_1335 import JOptSocAmB33_1335
 from .NewJPhys22_013049 import NewJPhys22_013049
 from .NewJPhys22_063041 import NewJPhys22_063041
 from .OptLett41_2676 import OptLett41_2676
+from .PhysRevA101_053836 import PhysRevA101_053836, PhysRevA101_053836_RWA
 from .PhysRevLett103_213603 import PhysRevLett103_213603
 from .PhysRevLett111_103605 import PhysRevLett111_103605
 from .PhysRevLett114_013601 import PhysRevLett114_013601